refactor(shared): Reorganize pagination classes into dedicated package
Structural changes (Shared)
- Created new `pagination` package (was in `dto`)
- Moved core pagination classes:
  * SimpleSlice (interface)
  * SimplePage (interface extends SimpleSlice)
  * Sliced (SimpleSlice implementation)
  * Paged (SimplePage implementation)

Cross-service updates:
- Adjusted all affected imports in:
  * gateway module
  * profile module

Contributes to 'profile' functionality
