+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/profile/geo/house/controller/PrivateHoseController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.pavbatol.myplace.gateway.app.access.RequiredRoles;
 import ru.pavbatol.myplace.gateway.app.api.ApiResponse;
 import ru.pavbatol.myplace.gateway.profile.geo.house.service.HouseService;
-import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.pagination.SimpleSlice;
 import ru.pavbatol.myplace.shared.dto.profile.geo.house.HouseDto;
 
 import javax.validation.constraints.Max;
