+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/profile/geo/street/controller/PrivateStreetController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.pavbatol.myplace.gateway.app.access.RequiredRoles;
 import ru.pavbatol.myplace.gateway.app.api.ApiResponse;
 import ru.pavbatol.myplace.gateway.profile.geo.street.service.StreetService;
-import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.pagination.SimpleSlice;
 import ru.pavbatol.myplace.shared.dto.profile.geo.street.StreetDto;
 
 import javax.validation.constraints.Max;
