+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/city/service/CityService.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.city.service;
 
-import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.pagination.SimpleSlice;
 import ru.pavbatol.myplace.shared.dto.profile.geo.city.CityDto;
 
 public interface CityService {
