+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/district/service/DistrictService.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.district.service;
 
-import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.pagination.SimpleSlice;
 import ru.pavbatol.myplace.shared.dto.profile.geo.district.DistrictDto;
 
 public interface DistrictService {
