+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/house/service/HouseService.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.house.service;
 
-import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.pagination.SimpleSlice;
 import ru.pavbatol.myplace.shared.dto.profile.geo.house.HouseDto;
 
 public interface HouseService {
