+++ mp-profile/src/main/java/ru/pavbatol/myplace/profile/service/ProfileService.java
@@ -2,7 +2,7 @@ package ru.pavbatol.myplace.profile.service;
 
 import ru.pavbatol.myplace.profile.dto.*;
 import ru.pavbatol.myplace.profile.model.ProfileStatus;
-import ru.pavbatol.myplace.shared.dto.pagination.SimplePage;
+import ru.pavbatol.myplace.shared.pagination.SimplePage;
 
 import java.util.UUID;
