+++ mp-shared/src/main/java/ru/pavbatol/myplace/shared/pagination/SimplePage.java
@@ -1,4 +1,4 @@
-package ru.pavbatol.myplace.shared.dto.pagination;
+package ru.pavbatol.myplace.shared.pagination;
 
 public interface SimplePage<T> extends SimpleSlice<T> {
     int getNumber();
