+++ mp-shared/src/main/java/ru/pavbatol/myplace/shared/pagination/SimpleSlice.java
@@ -1,4 +1,4 @@
-package ru.pavbatol.myplace.shared.dto.pagination;
+package ru.pavbatol.myplace.shared.pagination;
 
 import java.util.List;
